Core(cleanup): cleanup core header files
